requirements.txt
bardapi
requests
python-dotenv